moved tests in dedicated module
