+++ publib/__init__.py
@@ -4,3 +4,4 @@ publib
 """
 
 from .publib import set_style, fix_style
+from .test import *
