+++ publib/publib.py
@@ -8,16 +8,19 @@ For similar librairies, see seaborn, which also add neat high-end API to
 Matplotlib function calls.
 
 
+Description
 --------
-Use
+
+Use::
     
-    Call set_style() at the beginning of the script
-    Call fix_style() after each new axe is plotted
+    set_style()       # at the beginning of the script
+    ...
+    fix_style()       # after each new axe is plotted
 
 Note that importing publib will already load the basic style by default. 
 
-A couple more styles ('poster', 'article') can be selected with the function
-set_style()
+A couple more styles ('poster', 'article', 'origin') can be selected with the 
+function :func:`~publib.publib.set_style`
 
 Because some matplotlib parameters cannot be changed before the lines are 
 plotted, they are called through the function fix_style() which:
@@ -25,8 +28,8 @@ plotted, they are called through the function fix_style() which:
 - remove the spines
 - turn the legend draggable by default
 
---------
 Examples
+--------
 
     import numpy as np
     import matplotlib.pyplot as plt
@@ -41,13 +44,15 @@ Examples
     plt.show()
 
 
-Known issues
---------
+Notes
+-----
+
+Known issues:
 
 If fonts (ex: Times New Roman) appear to be bold, you may need Matplotlib to regenerate
 its font library cache (delete ~/.matplotlib/fontCache)
 
-See dedicated Stackoverflow:
+See dedicated Stackoverflow::
 
     https://stackoverflow.com/questions/33955900/matplotlib-times-new-roman-appears-bold
     
@@ -59,7 +64,6 @@ import matplotlib.pyplot as plt
 import matplotlib as mpl
 import os
 from os.path import dirname, join
-from warnings import warn
 from six import string_types
 
 style_params = {
@@ -112,6 +116,11 @@ def set_style(style='basic', **kwargs):
     >>> set_style('article')
 
     >>> set_style('poster',**{'lines.linewidth':2})
+    
+    See Also
+    --------
+    
+    :func:`~publib.publib.fix_style`
 
     '''
 
@@ -169,6 +178,11 @@ def fix_style(style='basic', ax=None, **kwargs):
     plb.set_style('poster')
     plt.plot(a,np.cos(a))
     plb.fix_style('poster',**{'draggable_legend':False})    
+    
+    See Also
+    --------
+    
+    :func:`~publib.publib.set_style`
 
     '''
 
@@ -274,107 +288,7 @@ def _get_lib():
 set_style('basic')        # whenever publib is imported
 
 
-# %% Test routines
-
-def _test(**kwargs):
-    ''' Note : for some reason using latex will also apply to the plot showing
-    the default behaviour of matplotlib (it seems that this parameter
-    is retroactive
-    '''
-
-    import numpy as np
-    import matplotlib.pyplot as plt
-
-    # %% Examples
-    def example1(title, seed):
-        np.random.seed(seed)
-        x = np.linspace(0, 5, 250)
-        y = np.cos(x)**2 + np.random.normal(scale=0.5, size=len(x))
-        yav = np.cos(x)**2
-        plt.figure()
-        ax = plt.subplot()
-        ax.plot(x, y, 'o', label='normal distribution')
-        ax.plot(x, yav, zorder=-1, label='average')
-        plt.xlabel(r'$x$')
-        plt.ylabel(r'$\cos^2 x$+noise')
-        plt.title(title)
-        plt.legend(loc='upper left')
-        plt.ylim((-1.5, 3.5))
-        plt.show()
-        return ax
-
-    def example2(title, seed):
-        np.random.seed(seed)
-        x = np.linspace(0, 5, 50)
-        y = np.cos(x)**2 + 1 * np.random.random(len(x))
-        yerr = np.std(y, axis=0)
-        plt.figure()
-        ax = plt.subplot()
-        ax.errorbar(x, y, yerr=yerr, fmt='o',
-                    label='new legends are draggable by default')
-        plt.xlabel(r'$x$')
-        plt.ylabel(r'$\cos^2 x$+noise')
-        plt.title(title)
-        plt.legend(loc='upper left')
-        plt.ylim((-1.5, 3.5))
-        plt.show()
-        return ax
-
-    def example3(title, seed):
-        np.random.seed(seed)
-        x = np.linspace(0, 5, 10)
-        y = np.cos(x)**2 + 0.1 * np.random.random(len(x))
-        yerr = np.std(y, axis=0)
-        plt.figure()
-        ax = plt.subplot()
-        ax.errorbar(x, y, yerr=yerr, marker='o', capsize=5,  # example of use of capsize
-                    label='scaled legend marker size')
-        plt.xlabel(r'$x$')
-        plt.ylabel(r'$\cos^2 x$+noise background')
-        plt.title(title)
-        plt.legend(markerscale=1.3)
-        plt.ylim((-1.5, 3.5))
-        plt.show()
-
-        return ax
-
-    # %% Plot them
-    plt.close('all')
-    import time
-
-    for example in [example1]:
-
-        seed = int(time.time())
-        mpl.rcdefaults()
-
-        set_style()
-        example('basic', seed)
-        fix_style()
-
-        set_style('article')
-        example('article', seed)
-        fix_style('article')
-
-#        set_style(['article','B&W'])
-#        example('article',seed)
-#        fix_style(['article','B&W'])
-
-        set_style('poster', **{'lines.linewidth': 5})
-        example('poster', seed)
-        fix_style('poster', **{'draggable_legend': False})
-
-        set_style(['origin'])
-        example('OriginPro', seed)
-        fix_style(['origin'])
-
-        # Default plot
-        mpl.style.use('classic')
-        example('matplotlib', seed)
-
-        mpl.style.use('ggplot')
-        example('ggplot', seed)
-
-    return True
 
 if __name__ == '__main__':
-    _test()
+    from test.test_functions import test_routines
+    test_routines()
