+++ publib/test/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""
+publib
+"""
+
+from __future__ import absolute_import
+
